Move to more modern versions of rails & ruby.
This deprecates support for Rails 3.x and 4.x.

Rails 3 is now more than three versions old.
Rails 4 doesn't play nicely with modern bundler.

This also deprecates support for Ruby 2.3, 2.4 and 2.5, which have either been end-of-lifed, or will be in March 2021.

This simplifies the travis file, and the Gemfile, hopefully allowing further upgrades to Ruby 3.0