Added support for Ruby 3.0
Ruby 3.0 deprecated allowing the last parameter of a method call being a hash, when using named arguments. This only occurred in one place, and was fixed.

However Rails 5.2.4.4 doesn't currently install properly with Ruby 3.0, while Rails 6 does work.

Given the previous commit removes support for a couple of versions of ruby, and this adds for new ones, and it's not really backwards compatible, it's probably worth bumping the major version.